reorganize segment files
